backend/requirements.docker.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
pymongo==4.13.2
PyJWT==2.10.1
pwdlib[argon2]==0.3.0
pydantic-settings==2.10.1
email-validator==2.2.0
python-multipart==0.0.20
pillow==11.3.0
google-genai==1.73.1
pytest==8.4.1
httpx==0.28.1
